[jak 2] Fix possible stereo desync in overlord (#2663)
Normally, when they allocate a VagCmd, they do a bunch of stuff to clear
all the status bits and reset things
in particular the InitVAGCmd function does a lot


![image](https://github.com/open-goal/jak-project/assets/48171810/9b355020-ad37-496c-9438-2f8d34f24e0a)

but for the stereo command, they do a lot less:

![image](https://github.com/open-goal/jak-project/assets/48171810/12a36712-0e68-4377-a6be-3bde82c2aa15)

Which means that the new_stereo_command can just have random status bits
left over from whatever the last user had.
we seem to end up in a state where byte21 is set, and this causes
everything else to be wrong and off-by-one dma transfer. My guess is
that the original game avoided this bug due to lucky timing that I don't
understand.

I think the fix of just clearing byte21 is ok because there's no way
that the old value of the byte is useful after the command is
repurposed.

+++ game/graphics/opengl_renderer/ocean/OceanTexture.cpp
@@ -75,7 +75,8 @@ void OceanTexture::init_textures(TexturePool& pool, GameVersion version) {
 
 void OceanTexture::draw_debug_window() {
   if (m_tex0_gpu) {
-    ImGui::Image((void*)m_tex0_gpu->gpu_textures.at(0).gl, ImVec2(m_tex0_gpu->w, m_tex0_gpu->h));
+    ImGui::Image((void*)(u64)m_tex0_gpu->gpu_textures.at(0).gl,
+                 ImVec2(m_tex0_gpu->w, m_tex0_gpu->h));
   }
 }
